core: set group info when creating new object

+++ core/src/main/java/org/usth/ict/ulake/core/persistence/GroupRepository.java
@@ -3,5 +3,5 @@ package org.usth.ict.ulake.core.persistence;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.usth.ict.ulake.core.model.LakeGroup;
 
-public interface GroupRepository extends JpaRepository<LakeGroup, Integer> {
+public interface GroupRepository extends JpaRepository<LakeGroup, Long> {
 }
